agent-ui/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
SQLAlchemy==2.0.35
passlib==1.7.4
python-jose==3.3.0
aiofiles==23.2.1
aiohttp==3.10.5
aiocache==0.12.2
pydantic==2.9.2
email-validator==2.1.1
requests==2.32.3
Jinja2==3.1.4
../genomeer # for dev only
